Move exampledata_utils out of tools and call tutorials
